requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
pytest==8.3.5
httpx==0.28.1
openai>=1.0.0
python-dotenv>=1.0.0
